доработал модуль генерации и вставки в разметку и логику вызова
