+++ js/main.js
@@ -1,1 +1,4 @@
-import './pictures.js'
+import { createOfferList } from './data.js';
+import { createPicturesElemts } from './create-pictures-elemtns.js';
+
+createPicturesElemts(createOfferList);
